refactor version handling
