+++ bin/oc-rsync/src/main.rs
@@ -1,8 +1,9 @@
 // bin/oc-rsync/src/main.rs
+mod version;
 use logging::LogFormat;
 use std::{io::ErrorKind, path::PathBuf};
 
-use oc_rsync_cli::{cli_command, parse_logging_flags, version_banner, EngineError};
+use oc_rsync_cli::{cli_command, parse_logging_flags, EngineError};
 use protocol::ExitCode;
 
 fn exit_code_from_error_kind(kind: clap::error::ErrorKind) -> ExitCode {
@@ -16,7 +17,7 @@ fn exit_code_from_error_kind(kind: clap::error::ErrorKind) -> ExitCode {
 fn main() {
     if std::env::args().any(|a| a == "--version" || a == "-V") {
         if !std::env::args().any(|a| a == "--quiet" || a == "-q") {
-            print!("{}", version_banner());
+            print!("{}", version::version_banner());
         }
         return;
     }
