+++ src/bin/version/mod.rs
@@ -1,0 +1,2 @@
+// src/bin/version/mod.rs
+include!(concat!(env!("CARGO_MANIFEST_DIR"), "/bin/oc-rsync/src/version.rs"));
